requirements.txt
anthropic>=0.25.0
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
rouge-score>=0.1.2
nltk>=3.8.1
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
PyPDF2>=3.0.1
python-docx>=1.1.0
beautifulsoup4>=4.12.0
tiktoken>=0.5.0
tabulate>=0.9.0
tqdm>=4.65.0
python-dotenv>=1.0.0
